+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## WIP
 ### Fixed
 - draw.io - convert HTML breaks `<br>` to new line characters.
+- PrefixingModder - improve regex match accuracy
 
 ### Added
 - allow defining render config in diagram